\name{clifford-package}
\alias{clifford-package}
\docType{package}
\title{
\packageTitle{clifford}
}
\description{
\packageDescription{clifford}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{clifford}
\packageIndices{clifford}

}
\author{
\packageAuthor{clifford}

Maintainer: \packageMaintainer{clifford}
}
\keyword{ package }
\seealso{\code{\link{clifford}}}
\examples{

rcliff()


}
